AI-Plants-System/requirements.txt
joblib==1.4.2
numpy==2.2.2
Pillow==11.1.0
streamlit==1.41.1
torch==2.6.0
transformers==4.48.1
